pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "diffsynth"
version = "2.0.7"
description = "Enjoy the magic of Diffusion models!"
authors = [{name = "ModelScope Team"}]
license = {text = "Apache-2.0"}
requires-python = ">=3.10.1"
dependencies = [
    "torch>=2.0.0",
    "torchvision",
    "transformers==5.4.0",
    "imageio",
    "imageio[ffmpeg]",
    "safetensors",
    "einops",
    "sentencepiece",
    "protobuf",
    "modelscope",
    "ftfy",
    "pandas",
    "pillow",
    "decord",
    "accelerate",
    "peft",
    "datasets",
    "opencv-python",
    "pytorch_lightning",
    "rerun_sdk",
    "tyro",
    "tqdm",
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
]

[tool.setuptools.packages.find]
where = ["./"]
include = ["diffsynth", "diffsynth.*"]

[project.optional-dependencies]
npu_aarch64 = [
    "torch==2.7.1",
    "torch-npu==2.7.1",
    "torchvision==0.22.1"
]
npu = [
    "torch==2.7.1+cpu",
    "torch-npu==2.7.1",
    "torchvision==0.22.1+cpu"
]
audio = [
    "torchaudio",
    "torchcodec"
]

[tool.setuptools]
include-package-data = true
